apps/admin-tool/requirements.txt
streamlit==1.31.0
opencv-python-headless
numpy
supabase
python-dotenv
Pillow
watchdog
streamlit-cropper
streamlit-drawable-canvas
setuptools
loguru
